+++ src/JitterDemo/Demos/Voxels/VoxelShape.cs
@@ -25,9 +25,9 @@ public class VoxelShape : RigidBodyShape
     public override void SupportMap(in JVector direction, out JVector result)
     {
         // this is the support function of a box with size 1.
-        result.X = Math.Sign(direction.X) * 0.5f;
-        result.Y = Math.Sign(direction.Y) * 0.5f;
-        result.Z = Math.Sign(direction.Z) * 0.5f;
+        result.X = Math.Sign(direction.X) * 0.5;
+        result.Y = Math.Sign(direction.Y) * 0.5;
+        result.Z = Math.Sign(direction.Z) * 0.5;
 
         result += Position;
     }
@@ -40,14 +40,14 @@ public class VoxelShape : RigidBodyShape
     public override void CalculateBoundingBox(in JQuaternion orientation, in JVector position, out JBBox box)
     {
         // NOTE: We do not support any transformation of the body here.
-        Debug.Assert(orientation.W > 0.999f, "Voxel shape can not be attached to a transformed body.");
+        Debug.Assert(orientation.W > 0.999, "Voxel shape can not be attached to a transformed body.");
         Debug.Assert(MathHelper.CloseToZero(position), "Voxel shape can not be attached to a transformed body.");
 
-        box.Min = Position - JVector.One * 0.5f;
-        box.Max = Position + JVector.One * 0.5f;
+        box.Min = Position - JVector.One * 0.5;
+        box.Max = Position + JVector.One * 0.5;
     }
 
-    public override void CalculateMassInertia(out JMatrix inertia, out JVector com, out float mass)
+    public override void CalculateMassInertia(out JMatrix inertia, out JVector com, out double mass)
     {
         // Do not try to calculate mass properties here.
         mass = 1;
